+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # UNRELEASED
 
+* EXOHostedContentFilterPolicy
+  * Fixed an issue with setting the default value for `IntraOrgFilterState`.
+    FIXES [#5031](https://github.com/microsoft/Microsoft365DSC/issues/5031)
 * DEPENDENCIES
   * Updated Microsoft.Graph to version 2.27.0.
   * Updated MSCoudLoginAssistant to version 1.1.43.